explode admin panel in independant components
