+++ src/shared/containers/admin/styles.js
@@ -1,0 +1,9 @@
+export const infoStyleD = {
+  fontSize: "16px",
+  fontWeight: "400",
+  color: "#666",
+  padding: "24px",
+  lineHeight: "1.1",
+};
+
+export const FORM_WIDTH = "100%";
